Save followed channels statuses inside NotifyApi
This fixes one type of errors when we fail to grab the *full* list of
followed channels and then we might miss the on<=>off switch. For
example:

=> Followed channel X is online
...
=> Channel X goes offline
...
=> Fail to get that the user follows channel X
=> Successfully grabbed all followed channels. (X is offline)
=> Successfully grabbed all followed channels. (X is offline)

In the old model we wouldn't have noticed the change because the two
most recent dictionaries are missing the actual change (they are both
offline). In this newer version we always have in mind the last status.
Thus when we get that X is offline and X is online in our class and then
we can show the notification.